test(fixtures): centralize complete runtime setup
