+++ tests/e2e/conftest.py
@@ -1,6 +1,3 @@
 """Complete-runtime fixtures used by the small public-surface workflows."""
 
-from tests.composition.conftest import (  # noqa: F401
-    authorized_complete_runtime,
-    authorized_portfolio_template,
-)
+pytest_plugins = ("tests.support.runtime_fixtures",)
